Refactor serverhub apps view helpers
